requirements.txt
# Core dependencies for AUD Daily Tracker
# Install with: pip install -r requirements.txt

# Data manipulation and analysis
pandas>=2.0.0

# API requests
requests>=2.31.0

# Date/time handling
python-dateutil>=2.8.2

# Timezone handling
pytz>=2023.3

# Excel file reading (for RBA historical data)
openpyxl>=3.1.0
xlrd>=2.0.1  # Required for reading .xls files (Excel 97-2003 format)

# HTML to image conversion (Selenium is more reliable than Playwright)
selenium>=4.15.0
playwright>=1.40.0  # Fallback option
Pillow>=10.0.0  # For PNG to JPEG conversion

# Web scraping (TradingEconomics)
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0

